SPEAr/db9000: pass size aligned to section for memblock alloc
db9000 clcd driver allocates memory at boot time for its usage aligned
to 1MB boundary but it freed actual size of memory it asked for.

This created a problem with DT as memory blocks created as a result of
db9000 calling 'free' was not section (1 MB) aligned.

The memblock blocks can be aligned to section size if we allocate
and free all memories aligned to section boundary.

To ensure this pass size aligned to section size before calling memblock
alloc. Also use 'arm_memblock_steal' instead which is a better wrapper to
call a sequence of memblock alloc, reserve and free.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>