stmmac: Remove memory leaks
This patch removes the memory leaks associated at the time of
registration. At present since the platform data was passed from
platform as Aux data, and the DT was being used to pass the IO address
and IRQ numbers, both the conditions for DT check and platform data were
true.

The code in DT subsection was allocating memory which was not being
used. This patch corrects the same leakage from the code.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>