char/spear-adc : Fix for ADC data transfers using DMA.
In the recent alignment of the DMA driver, the driver requires the
dma_address field in the scatterlist to be filled with physical address
of the buffer to be used for transfers.

This patch fills in the physical address of the buffer in the
dma_address field of the scatterlist instead of the virtual address.

Previously the DMA driver was not using this and hence forth the bug did
not came up.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>