ARM: abma: Dis/enable interface clock in PM
amba_probe enables amba device interface clock and always leaves it enabled on
successful probe of amba device. In conventional PM we need to disable this
interface clock and re-enable it while resuming system. This is what it handles
in runtime PM.

This patch is to disable interface clock for amba devices from amba layer.

Signed-off-by: Vipul Kumar Samar <vipulkumar.samar@st.com>