ARM:dts:SPEAr1340: Re-arrange sound card entry.
Re-arrange sound card entry in spear1340-evb file.

It shows aplay and arecord device listing as

  $ alay -l
  card 0: spearevb [spear-evb], device 1: spdif-play dit-hifi-1 []

and

  $ arecord -l
  card 0: spearevb [spear-evb], device 0: spdif-cap dir-hifi-0 []

which doesn't match with previous release lsp-3.2.5.

Fix device card number under aplay and arecord card listing to be
compatible with previous release order.

With this patch, the output for the aplay-l and arecord -l is

  $ alay -l
  card 0: spearevb [spear-evb], device 0: spdif-play dit-hifi-1 []

and

  $ arecord -l
  card 0: spearevb [spear-evb], device 1: spdif-cap dir-hifi-0 []

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>