Clk:SPEAr: Modify CPU clock related configurations
This patch adds in the following support:

1. For platform specific implementation of CPU Freq on SPEAr
6xx/3xx/1310, the changes in CPU clock are to be iterated up the
hierarchy to VCO/PLL to ensure that the CPU frequency is altered.

This requires CLK_SET_RATE_PARENT flag to be set while the cpu clock is
being registred. This patch sets in the same.

2. Additionally for SPEAr6xx and SPEAr3xx, the PLL table value is
restricted to a single entry i.e. p value of 1. This value is sufficient
to generate the requisite fequencies of 166/266/332 Mhz. These
frequencies are stable system frequencies.

---------------------------
LIMITATION:
(Catering to point number-2)
----------------------------

Since the DDR shares the same pll as the total system (pll-1), the
changes in system frequency requires the DDR to be moved into self
refresh mode.

The Clock framework implemetations changes VCO and PLL separately to the
m/n & p values maintained in the VCO and PLL rate tables respectively.

A quick fix by limiting the p-value to 1 helps to overcome the above
limitations and handle the DDR specific changes at VCO level only.

Please note that since the pll is shared with system additions or
deletions into the vco and pll rate tables is subject to system
unstability and would required thorough integration tests.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>