SPEAr6xx/3xx: clk: Handle DDR maintainence while setting VCO
On SPEAr6xx and SPEAr3xx machines, the DDR is being run on the same
PLL's as is the entire system.

On change of request of CPU frequency by the CPU freq framework, the PLL
frequency could not be altered straight away without taking into the
account the fact that the system memory also shares the same PLL.

This patch explicitly moves the DDR into self refresh mode, and changes
system mode to slow before changing in the VCO parameters, and then
restores the system to normal mode followed by transition of DDR out of
self refresh mode.

The parameters changes for VCO only involves m(PLL feedback divisor) and
n (PLL pre divisor) value configurations.

For the cases, when the DDR runs on a separate PLL then system, the
change of system frequency will not impact the memory and normal config
changes could be carried out at VCO/PLL level.

WARNING:
Since the change of system frequency is altering the memory input clock,
this change may render instability to the system.

The tests have been performed on 166/266/332 Mhz, and these frequencies
are tested to be stable for the system.

The p-value(PLL post divisor) value for these system frequencies is
always 1, hence the routine in this patch has been called at VCO level.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>