SPEAr 6xx/3xx: Create static non-cached memory mapping for SRAM
This patch creates 4K of static non-cached memory mapping for SRAM
required for storing and executing the suspend to ram code.

The low level handlers for suspend to ram rely on the static addresses
for latching few registers and values.

This mapping would be removed in future, as we would cater to the low
level suspend to ram code in more efficient way.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>