SPEAr3xx/6xx: PM: Pass the physical address for cpu resume.
The low level handler for suspend to ram platform latches the physical
address of cpu_resume() before moving to sleep and then restores and
jump to this address on resume.

This patch rectifies the virtual address of cpu_resume being passed to
the low level handler and now correctly passes the physical address for
the same.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>