video:db9000:Fix: Initialize backlight in open callback
In current implementation we enable db9000_controller in open callback and
configure backlight in probe. In backlight configuration we set default
backlight brightness to the value written in pwmdcr(duty cycle register).

But at the time of probe controller is disable so register is not accessible and
it configure default brightness to zero. Because of this reason while playing
video on X server with suspend/resume test brightness reset to zero.

Signed-off-by: Vipul Kumar Samar <vipulkumar.samar@st.com>