pinctrl:SPEAr1310: Separate out pci pins from pcie_sata pin group
SPEAr1310 has separate PCI and PCIe implementations which are not muxed
with each other. Presently they have been implemented as muxed together
with SATA and represented wrongly in the software.

In reality only PCIe and SATA implementations are muxed with each other.
This patch separates out pci pins creating a new pingroup and function
for pci.

Signed-off-by: Vipul Kumar Samar <vipulkumar.samar@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>