SPEAr: Add JPEG device platform support
This patch provides the DMA Slave related configurations for JPEG from
the platform using AUX data.

This is subject to change in future, as AUX data would be removed.

NOTE:
For SPEAr1310, the specific dma callback for jpeg has been defined to
make some flow controller related settings related to DMA.

Additionally, the JPEG driver requires reset at the HW level which is
controlled from Misc space. This control has been added as a callback in
the platform data.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>
Reviewed-by: Vipul Kumar Samar <vipulkumar.samar@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>