can: c_can: Correct register alignment info passing through DT
Depending upon the integration of the C_CAN controller in the SoC, the
registers may be aligned to 16-bit or 32-bit boundary. The device tree
support added to the driver assumes this information can be passed
through 'flags' in the platform resources, however current DT
implementation provides no such provision.

So passing the information explicitly through the optional parameter
'reg_alignment' in the device tree.

Signed-off-by: Amit Virdi <amit.virdi@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>