SPEAr: cpufreq: Fix for symlinks on hotplug
For the secondary cores that could be hot plugged on run time, the
symlinks created during bootup related to cpu freq policy were not
observed in the sysfs interface. This behavior was seen as and when
cpu-1 was made offline and then online again.

This patch provides a fix for the same, and additionally limits the mask
settings for policy related to cpufreq to smp cores.

During the bootup, primary cpu core manages the cpufreq interface with
mask (policy->cpus) being set for all online cpus, and (policy->cpu) set
to 0.

Because of the above mask settings the checks that existed in the spear
cpu freq driver (policy->cpu != 0) allowed the transitions in the cpu
freq framework from all cpu interfaces post bootup.

During the hotplug event, once the cpu is made offline and online again,
a new policy is allocated with policy->cpu set to core number(1), and
the init callback of the spear cpufreq driver is executed.

The above check done in the spear cpu freq driver failed and hence the
symlinks were not created

In case of success, post this the cpufreq core deallocates the policy,
since the mask being set in the init code ensures that cpu0 will manage
the policy and symlinks from rest of cores would be retained.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>