clk/SPEAr1310: Enable ras_ahb_clk permanently
This commit modifies the following:

	commit fcf3e5bec35c7f160af05a8d6cd6399e083ab194
	Author: Amit Virdi <amit.virdi@st.com>
	Date:   Thu Nov 8 10:45:30 2012 +0530

	clk/SPEAr1310: Don't disable ras_ahb_clk

	CAN requires ras_apb_clk and ras_ahb_clk for it's operation. The
	current clock framework hierarchy provides option to enable only one
	clock. So use flag to ignore disabling ras_ahb_clk

The above commit had a limitation that when a RAS IP is disabled, then
ras_ahb_clk is also gated as the usage count becomes zero. Enabling
ras_ahb_clk using the new implementation increments it's usage count and
hence, even if no RAS IP is in use, this clock remains enable.

This modification has been tested on SPEAr1310 EVB.

Signed-off-by: Amit Virdi <amit.virdi@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>