input:stmpe-ts: Add a delay before resetting fifo
Issue:
On long time pressing touchscreen(TS) or some time after few touches the TS
driver stops.

Observation:
1) At this point if we dump stmpe-ts register most of the time it shows that there
is a touch event interrupt pending and SPEAr does not serve it.
If we reset the fifo and clear all interrupts, TS work again.

2) Other observation is if we configure i2c in slow mode TS working fine.

So my conclusion is that due to some timing issues or speed mismatch SPEAr is
not able to serve all the interrupts.

Solution:
Introduce a delay of 10us in between resetting fifo so that i2c get sufficient
time to serve all the interrupts.

Signed-off-by: Vipul Kumar Samar <vipulkumar.samar@st.com>