SPEAr1310 PM: Workaround for wake from ethernet in s2r mode
This patch provides a software workaround for a bug in  bootrom, where
the ethernet wake up trigger is not detected because of the flaw in the
mask used to read the wake up register.

Presently, bootrom wrongly tries to mask for possible wakeup sources
as (0x1F << 5) in PCM Configuration Register (offset 0x100 of Misc
space). This mask should be (0x3F << 6) for SPEAr1310.

As a result of this bootrom misses Ethernet wakeup event (bit 10 of PCM
CFG). As a workaround targeted by this patch, PCIE wakeup enable bit is
set to permanent high thus indicating wakeup event (as it falls within
the mask being used in the bootrom), and helps to bring the system out
of suspend to ram when ether wake occurs.

It would not affect the normal bootflow where PCM CFG wakeup events
would be all cleared due to SoC reset.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>