usb/designware_udc: clear all interrupts in device irq handler
It is observed that during bus suspend a bit in reserved part of device irq
register is set which keeps interrupting cpu. Presently interrupt handler
clears only those condition which it handles, thus giving a chance to this
erroneous condition to persist.

As a workaround clear all bits set in the device interrupt register.

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>