ASoC:Codec:STA529: Correct S2PCFG0 register bits for play.
Setting BICLK_STRB in S2PCFG0 register. It means that bit clock strobe
edge is rising edge, bit clock active edge is falling edge. It indicates
that first data will be place on the bus, and then pulse the strobe line
to instruct device that the data is ready.

Earlier value for this register was 0xb2. For capture on SPEAr1340,
since sta529 record part operates in master mode, bit clock would not be
shared for play and record lines of sta529(bit number 5 at offset 0xa).
So this register value must be 0x92 instead  of 0x12.

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>