video:db9000:Fix: Correct the logic for pixel clock programming.
This patch corrects the logic for pixel clock programming. Previously the
preferred way was to use internal db9000 divider to generate the pixel clock,
resulting in inaccurate frequency dividers.  Use of SSCG results in more
precise pixel clock output. A check is made to be sure we are in the 0.5%
tolerance of VESA standard.

Signed-off-by: Salvatore De Dominicis <salvatore.dedominicis@st.com>