plat-spear/time.c: timer asynchronous issue workaround
This patch fixes a possible issue related to the fact that
in the GPT timer IP the two basic clocks, the bus_clk (pclk)
to read the counter and the timer_clk to increment it, are
asynchronous to each other. When reading it may very rarely
happen to get an inconsistent value, as read_op may enter
during a timer counter transition.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>